# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des utilisateurs et vendeurs.",
        "title": "o'z api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support API",
            "email": "goliyaohugues@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "summary": "Liste paginée des utilisateurs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Numéro de la page (défaut: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nombre d'éléments (défaut: 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permet à un Admin ou Super Admin de créer un utilisateur.\nSi l'auteur est Admin, le rôle est forcé à 'vendeur' et lié à son ID.\nSi l'auteur est Super Admin, il peut définir le rôle.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Créer un nouvel utilisateur (Admin/Vendeur)",
                "parameters": [
                    {
                        "description": "Données de l'utilisateur à créer (email, username, password, firstname, lastname, phone)",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ex: {message: '...', user: {models.User}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Données JSON invalides ou erreur de validation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Accès refusé : privilèges insuffisants",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Récupère les détails d'un utilisateur par son UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Récupère un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Format d'identifiant invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Aucun utilisateur trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Met à jour les informations. Seul 'super' peut changer le rôle ou le parent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Met à jour un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transport.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Erreur de permission",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Supprime un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Restaure un utilisateur",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Utilisateur restauré",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "password",
                "phone",
                "role",
                "username"
            ],
            "properties": {
                "UUID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "parentID": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                },
                "profilePicture": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "transport.UpdateUserRequest": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "nouveau@email.com"
                },
                "firstname": {
                    "type": "string",
                    "example": "Prince"
                },
                "lastname": {
                    "type": "string",
                    "example": "Konan"
                },
                "parentId": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string",
                    "example": "0102030405"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "vendeur"
                    ],
                    "example": "vendeur"
                },
                "username": {
                    "type": "string",
                    "example": "nouveau_pseudo"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" suivi d'un espace et du token JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}